requirements.txt
kornia
numpy
opencv-python-headless
Pillow
PyYAML
requests
scikit-image
scipy
tensorboard-pytorch
torch
torchvision
tqdm
wandb
ipython
tensorboard
statsmodels
pytorch-fid
imgaug
scikit-learn
flask
pycocotools
csaps
